requirements_new.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
fastjsonschema==2.15.3
filelock==3.6.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.7.2
google-api-python-client==2.45.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
httplib2==0.20.4
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
ipykernel==6.13.0
ipython
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
metrics
mistune==0.8.4
nbclient==0.6.0
nbconvert
nbformat==5.3.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.11
numpy==1.21.6
oauthlib==3.2.0
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pathlib2==2.3.7.post1
pathspec==0.5.5
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.7.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
Pygments
Pympler==1.0.1
pyparsing==3.0.8
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.5
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
sacremoses==0.0.49
scikit-learn==1.0.2
scipy==1.7.3
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
stopwords==1.0.0
streamlit==1.8.1
streamlit-lottie==0.0.3
tenacity==8.0.1
terminado==0.13.3
textblob==0.17.1
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
toolz==0.11.2
torch==1.11.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.18.0
tweepy==4.8.0
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.9
validators==0.18.2
watchdog==2.1.7
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.6.0
zipp==3.8.0
